chore: add heroku config settings to enable hosting on heroku
