+++ package.json
@@ -8,7 +8,8 @@
     "test:db": "set NODE_ENV=test&&sequelize db:migrate:undo:all&&sequelize db:migrate && cross-env NODE_ENV=test nyc mocha server/test/test.js",
     "migrate": "sequelize db:migrate:undo:all&&sequelize db:migrate",
     "start:dev": "nodemon --exec babel-node server/bin/www",
-    "coverage": "nyc report --reporter=text-lcov | coveralls"
+    "coverage": "nyc report --reporter=text-lcov | coveralls",
+    "heroku-postbuild": "sequelize db:migrate"
   },
   "nyc": {
     "require": [
@@ -34,13 +35,13 @@
   "homepage": "https://github.com/Iyke09/Events#readme",
   "dependencies": {
     "babel-cli": "^6.26.0",
-    "babel-core": "^6.25.0",
+    "babel-core": "^6.25.0",//----------------------
     "babel-istanbul": "^0.12.2",
-    "babel-loader": "^7.1.1",
+    "babel-loader": "^7.1.1",//-----------------------
     "babel-plugin-istanbul": "^4.1.4",
-    "babel-polyfill": "^6.23.0",
+    "babel-polyfill": "^6.23.0",//------------------------
     "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-es2015": "^6.24.1",//------------------------
     "babel-preset-latest": "^6.22.0",
     "babel-preset-react": "^6.24.1",
     "babel-register": "^6.26.0",
